ENV var

+++ lib/RecordsKeeperRubyLib/address.rb
@@ -12,13 +12,22 @@ require 'yaml'
 # Address class to access address related functions
 module RecordsKeeperRubyLib
 	class Address
-		# Import values from configuration file.
-		cfg = YAML::load(File.open('config.yaml','r'))
-		@network = cfg['network']
-		@url = cfg['network']['url']
-		@user = cfg['network']['rkuser']
-		@password = cfg['network']['passwd']
-		@chain = cfg['network']['chain']
+		if File.exist?('config.yaml')
+			# Import values from configuration file.
+			cfg = YAML::load(File.open('config.yaml','r'))
+			@network = cfg['network']
+			@url = cfg['network']['url']
+			@user = cfg['network']['rkuser']
+			@password = cfg['network']['passwd']
+			@chain = cfg['network']['chain']
+		else
+			#pp ENV
+			@network = ENV['network']
+			@url = ENV['url']
+    		@user = ENV['rkuser']
+    		@password = ENV['passwd']
+    		@chain = ENV['chain']	
+		end
 
 		def self.variable
 			net = @network
@@ -40,6 +49,7 @@ module RecordsKeeperRubyLib
 			address = out[0]['result']
 			return address
 		end
+		puts getAddress
 
 		# Function to generate a new multisignature address
 		def self.getMultisigAddress nrequired, key		#getMultisigAddress() function definition
